Applied some of patch [ 688466 ] MSVC7 build & bug fixes by Darren Whobrey
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@19930 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/notebook.h
@@ -67,7 +67,7 @@ public:
     // ---------
 
     // get number of pages in the dialog
-    int GetPageCount() const { return m_pages.GetCount(); }
+    int GetPageCount() const { return (int) m_pages.GetCount(); }
 
     // get the panel which represents the given page
     wxNotebookPage *GetPage(int nPage) { return m_pages[nPage]; }
